src/main/java/org/study/parksangkhil/OddEvenLinkedList.java
package org.study.parksangkhil;

public class OddEvenLinkedList {
    public ListNode solution(ListNode head) {
        if (head == null) return null;

        ListNode odd = head;
        ListNode even = head.next;

        ListNode evenHead = even;

        while (even != null && even.next != null) {
            odd.next = odd.next.next;
            even.next = even.next.next;

            odd = odd.next;
            even = even.next;
        }

        odd.next = evenHead;

        return head;
    }
}